requirements.txt
beautifulsoup4==4.14.3
cloudscraper==1.2.71
requests==2.32.5
selenium==4.40.0
google-cloud-storage==3.9.0

google-api-core==2.29.0
google-auth==2.48.0
google-cloud==core-2.5.0
google-cloud==storage-3.9.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common==protos-1.72.0
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1-modules==0.4.2
rsa==4.9.1
